feat(v1.0): 플레이 시간 소수점 2자리 + 구글 캘린더 655시간 임포트
캘린더 세션 기록이 15분 단위라 합계가 정수로 안 떨어진다
(1시간 45분 = 1.75시간). integer로 두면 반올림하는 순간 원본이 사라진다.

V4: play_time_hours integer -> numeric(7,2)
- BacklogEntry.playTimeHours Integer -> BigDecimal, setScale(2) + 범위 검증
- PersonalRecord/PersonalRecordRequest/DeletedEntryDetail/MemberExport DTO
- StatsQueryService sumLong -> sumBigDecimal, PlaytimeStats.totalHours BigDecimal
- 프론트: lib/format.ts 신설(formatHours/roundHours), 입력칸 소수점 허용

캘린더 임포트 (04. 개인 (게임), 2025-12-15~2026-08-23)
- 329 세션 / 88개 이름 -> 35개 게임 654.99시간
- 전 항목이 비어 있던 칸이라 덮어쓴 값 없음
- 백업: backups/calendar/ (gitignore)

테스트 455/456 (PostgresSchemaTest는 Docker 미기동)
계약 검사 26건 불일치 없음

+++ backend/src/main/java/com/milobeene/starlog/backlog/dto/BacklogDetailResponse.java
@@ -109,7 +109,7 @@ public record BacklogDetailResponse(
             MoneyResponse listPrice
     ) {}
 
-    public record PersonalRecord(BigDecimal rating, Integer playTimeHours, String memo) {}
+    public record PersonalRecord(BigDecimal rating, BigDecimal playTimeHours, String memo) {}
 
     /** name은 "거실 스위치 (Nintendo Switch)" 꼴이다 — 선택지 목록과 같은 문구여야 짝이 맞는다 */
     public record DeviceRef(Long deviceId, String name) {}
