+++ backend/src/main/java/com/milobeene/starlog/backlog/dto/PersonalRecordRequest.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
  */
 public record PersonalRecordRequest(
         BigDecimal rating,          // 0.0~100.0 범위는 엔티티가 본다 (도메인 불변식)
-        Integer playTimeHours,      // 음수 금지도 엔티티가 본다
+        BigDecimal playTimeHours,      // 음수 금지도 엔티티가 본다
         /*
          * TEXT 컬럼이라 도메인 제한이 아니라 폭주 방지 상한이다.
          * 2000 → 5000: 옵시디언에서 옮겨온 감상 하나가 3166자였다. 상한 때문에 사용자가
